pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "drift"
version = "0.1.0"
description = "Local Python CLI for deterministic + AI-assisted MNQ trade planning."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.7,<3",
  "python-dotenv>=1.0,<2",
  "PyYAML>=6.0,<7",
  "rich>=13.7,<15",
  "typer>=0.12,<1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2,<9",
  "pytest-cov>=5,<7",
  "ruff>=0.5,<0.12",
  "mypy>=1.10,<2",
]

[project.scripts]
drift = "drift.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

